requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
cssselect2==0.7.0
daphne==4.1.2
diff-match-patch==20241021
dj-database-url==2.3.0
Django==4.2.18
django-import-export==4.1.1
django-jazzmin==3.0.1
fonttools==4.57.0
gunicorn==23.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.4.2
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
pyOpenSSL==25.0.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
scikit-surprise==1.1.4
scipy==1.10.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
tablib==3.5.0
tinycss2==1.4.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
weasyprint==61.2
webencodings==0.5.1
zope.interface==7.2
zopfli==0.2.3.post1
